+++ web/html/src/global.d.ts
@@ -1,6 +1,18 @@
 declare global {
   interface Window {
     _IS_UYUNI: boolean;
+
+    // Translation data
+    translationData?: any;
+    preferredLocale?: string;
+  }
+
+  // WIP test env setup, see ./utils/test-utils
+  namespace NodeJS {
+    interface Global {
+      Loggerhead: any;
+      jQuery: (window: Window, noGlobal?: boolean) => JQueryStatic;
+    }
   }
 
   function t(msg: string, ...args: Array<any>): string;
