+++ web/html/src/utils/data-providers/types.ts
@@ -1,8 +1,6 @@
-// @flow
-
 export type PagedData = {
-  items: Array<any>,
-  total: number
+    items: Array<any>;
+    total: number;
 };
 
 export type Comparator = (a: any, b: any, key: string, direction: number) => number;
